syntax refactored
