+++ src/presenter/main-presenter.js
@@ -1,11 +1,10 @@
 import { render } from '../render';
-import AddFormView from '../view/add-form-view';
-// import EditFormView from '../view/edit-point-view';
+import EditFormView from '../view/edit-form-view';
 import FiltersView from '../view/filters-view';
 import PointView from '../view/point-view';
 import PointsListView from '../view/points-list-view';
 import SortView from '../view/sort-view';
-import { POINTS } from '../data';
+import { POINTS } from '../const';
 
 // $======================== MainPresenter ========================$ //
 
@@ -20,8 +19,8 @@ export default class MainPresenter {
   init() {
     render(new FiltersView(), this.filtersContainer);
     render(new SortView(), this.tripEventsContainer);
-    render(new AddFormView(), this.tripEventsContainer);
+    render(new EditFormView(), this.tripEventsContainer);
     render(this.pointsListElement, this.tripEventsContainer);
-    POINTS.map((point) => render(new PointView(point), this.pointsListElement.getElement()));
+    POINTS.forEach((point) => render(new PointView(point), this.pointsListElement.getElement()));
   }
 }
